fixup(uwu): get remote endpoint in client

+++ courses/netdev/theories/uwu/Library/NetworkConfiguration.cs
@@ -2,5 +2,5 @@ namespace uwu.Library;
 
 public static class NetworkConfiguration
 {
-    public const short DEFAULT_PORT = 5667; 
+    public const int DEFAULT_PORT = 5667; 
 }
